Rename {opt} arguments as {config}
I found this part of the documentation confusing the first time I read
it. Since `opt` has a particular meaning in the context of Vim's
packages, I assumed that the `{opt}` argument had something to do with
`opt` plugins. I think it's clearer to use a different name for the
dictionary.