Implement progress window
Outputting the log messages by echo commands is sometimes not good
especially when updating many plugins. It may mess the display. Add a
new option to output the log message to a separate window, and enable it
by default.

This adds or changes the following options in `minpac#init()`:

* progress_open: Open the progress window if set to 'horizontal',
  'vertical' or 'tab'.
* status_open: Changed the default from 'vertical' to 'horizontal'.
* status_auto: Open the status window automatically if set to v:true.

Note that these options change the default behavior from minpac v2.0.x
or earlier.