Move plugin/minpac.vim to autoload/
The original intention of using `plugin/minpac.vim` was to reduce the
files to be loaded as much as possible, but the mandatory file to be
always loaded.  However, putting autoload-style functions into `plugin`
is tricky, and this might not be effective as I thought.
Use `autoload` instead of `plugin`.